Initial logic and separate out context
