+++ packages/datatrak-web/src/features/Survey/SurveyContext/index.ts
@@ -1,0 +1,6 @@
+/*
+ * Tupaia
+ *  Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
+ */
+
+export { SurveyContext, useSurveyForm } from './SurveyContext';
